@@ -19,7 +19,7 @@ public struct ScreenTopAction<
         navigator.logInfo("Searching for a top container of \(Output.self) type in \(type(of: container))")
 
         let top = navigator.topContainer(in: container) { container in
-            self.predicate.checkContainer(container)
+            predicate.checkContainer(container)
         }
 
         guard let output = top as? Output else {
