@@ -52,7 +52,7 @@ public final class DefaultScreenObservatory: ScreenObservatory {
         updateObservers(appending: storage)
 
         return ScreenObserverToken { [weak self, weak storage] in
-            if let self = self, let storage = storage {
+            if let self, let storage {
                 self.updateObservers(removing: storage)
             }
         }
