@@ -122,9 +122,9 @@ extension URLQuerySingleValueDecodingContainer: Decoder {
     }
 }
 
-private extension DecodingError {
+extension DecodingError {
 
-    static func keyedContainerTypeMismatch(
+    fileprivate static func keyedContainerTypeMismatch(
         at codingPath: [CodingKey],
         component: URLQueryComponent?
     ) -> Self {
@@ -147,7 +147,7 @@ private extension DecodingError {
         return .typeMismatch([String: Any].self, Context(codingPath: codingPath, debugDescription: debugDescription))
     }
 
-    static func unkeyedContainerTypeMismatch(
+    fileprivate static func unkeyedContainerTypeMismatch(
         at codingPath: [CodingKey],
         component: URLQueryComponent?
     ) -> Self {
