@@ -763,7 +763,7 @@ final class URLQueryDecoderTests: XCTestCase, URLQueryDecoderTesting {
             }
 
             override func isEqual(_ object: Any?) -> Bool {
-                guard let object = object as? DecodableSubclass else {
+                guard let object = object as? Self else {
                     return false
                 }
 
